+++ lib/redis_failover/node_watcher.rb
@@ -1,6 +1,8 @@
 module RedisFailover
-  # Watches a specific redis node for its reachability.
+  # Watches a specific redis node for its availability.
   class NodeWatcher
+    WATCHER_SLEEP_TIME = 3
+
     def initialize(manager, node, max_failures)
       @manager = manager
       @node = node
@@ -27,19 +29,23 @@ module RedisFailover
     def monitor_node
       failures = 0
 
-      begin
-        return if @done
-        @node.ping
-        failures = 0
-        @manager.notify_state_change(@node, :reachable)
-        @node.wait_until_unreachable
-      rescue NodeUnreachableError
-        failures += 1
-        if failures >= @max_failures
-          @manager.notify_state_change(@node, :unreachable)
+      loop do
+        begin
+          return if @done
+          sleep(WATCHER_SLEEP_TIME)
           failures = 0
+
+          if @node.available?
+            @manager.notify_state_change(@node, :available)
+            @node.wait_until_unavailable
+          end
+        rescue NodeUnavailableError
+          failures += 1
+          if failures >= @max_failures
+            @manager.notify_state_change(@node, :unavailable)
+            failures = 0
+          end
         end
-        sleep(3) && retry
       end
     end
   end