+++ spec/node_spec.rb
@@ -20,13 +20,13 @@ module RedisFailover
     end
 
     describe '#ping' do
-      it 'responds properly if node is reachable' do
+      it 'responds properly if node is available' do
         expect { node.ping }.to_not raise_error
       end
 
-      it 'responds properly if node is unreachable' do
-        node.redis.make_unreachable!
-        expect { node.ping }.to raise_error(NodeUnreachableError)
+      it 'responds properly if node is unavailable' do
+        node.redis.make_unavailable!
+        expect { node.ping }.to raise_error(NodeUnavailableError)
       end
     end
 
@@ -53,18 +53,18 @@ module RedisFailover
       end
     end
 
-    describe '#wait_until_unreachable' do
+    describe '#wait_until_unavailable' do
       it 'should wait until node dies' do
-        thread = Thread.new { node.wait_until_unreachable }
+        thread = Thread.new { node.wait_until_unavailable }
         thread.should be_alive
-        node.redis.make_unreachable!
+        node.redis.make_unavailable!
         expect { thread.value }.to raise_error
       end
     end
 
     describe '#stop_waiting' do
       it 'should gracefully stop waiting' do
-        thread = Thread.new { node.wait_until_unreachable }
+        thread = Thread.new { node.wait_until_unavailable }
         thread.should be_alive
         node.stop_waiting
         sleep 0.2