+++ Changes.md
@@ -1,12 +1,13 @@
 HEAD
 -----------
-- No longer force newly reachable slaves to master if already slaves of that master
+- No longer force newly available slaves to master if already slaves of that master
+- Honor a node's slave-serve-stale-data configuration option; do not mark a sync-with-master-in-progress slave as available if its slave-serve-stale-data is disabled)
+- Change reachable/unreachable wording to available/unavailable
 
 0.3.0
 -----------
 - Integrated travis-ci
-- Added background monitor to client for proactively detecting
-  changes to current set of redis nodes
+- Added background monitor to client for proactively detecting changes to current set of redis nodes
 
 0.2.0
 -----------